Refactor NodePiece into several modules
